requirements.txt
aniso8601==4.1.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
geoip2==2.9.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
maxminddb==1.4.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
urllib3==1.24.1
Werkzeug==0.14.1
